Revert "Add TestDetailsDialog component"
This reverts commit 25beb4e2183c5351f498cfc48754ef5144d55951.
